Change name to Ark
Ark was the orginal name that I wanted, because of the multiple backups.
I envisioned Noah's ark with two of each animal. I was worried about
conflict with the Arq backup program. However, I read the "about" page
on the Arq web site and it does not mention anything about Noah's Ark
being a motivation. And the name just fits perfectly. So I am switching
back to Ark!