add placeholder ark logo
Logo from https://openclipart.org/detail/193158/noahs-ark

Logo license:

We use the Creative Commons Zero 1.0 Public Domain License every time an
artist uploads a piece of clipart to Openclipart to make it clear the
artist is releasing the creative work for anyone to use for any reason,
even commercially. This act of "sharing" is the foundation Openclipart
is based upon. Sometimes people want a disclaimer or legal assurance
that our clipart is safe to be used. This extra security request often
comes from people working for companies of all sizes and, or,
publishers. If you need this type of document, please visit, save,
download and, or print the following for your needs: Creative Commons
Zero 1.0 License.